+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added new command `jobstat`, which provides system-wide view of application
+  status data.
 - Added a `sat bootprep list-vars` subcommand which lists the variables
   available in bootprep input files at runtime.